refactor: modularize statistical data point creation and update interfaces for better structure
