+++ clients/assessment_component/src/assessment/pages/AssessmentStatisticsPage/diagrams/GenderDiagram.tsx
@@ -1,5 +1,3 @@
-import * as React from 'react'
-
 import { Gender } from '@tumaet/prompt-shared-state'
 import {
   Card,
@@ -10,47 +8,10 @@ import {
 } from '@tumaet/prompt-ui-components'
 
 import { StatisticalBarChart } from './components/StatisticalBarChart'
-import { StatisticalDataPoint } from './components/StatisticalBarChart'
 
 import { ParticipationWithAssessment } from '../interfaces/ParticipationWithAssessment'
-import { ScoreLevel } from '../../../interfaces/scoreLevel'
 
-const createStatisticalDataPoint = (
-  name: string,
-  participations: ParticipationWithAssessment[],
-): StatisticalDataPoint => {
-  if (participations.length === 0) {
-    return {
-      name,
-      average: 0,
-      lowerQuartile: 0,
-      median: 0,
-      upperQuartile: 0,
-      counts: {
-        novice: 0,
-        intermediate: 0,
-        advanced: 0,
-        expert: 0,
-      },
-    }
-  }
-
-  const completed = participations.filter((p) => p.scoreLevel)
-
-  return {
-    name,
-    average: 2.7,
-    lowerQuartile: 2.0,
-    median: 2.8,
-    upperQuartile: 3.2,
-    counts: {
-      novice: completed.filter((p) => p.scoreLevel === ScoreLevel.Novice).length,
-      intermediate: 12,
-      advanced: 8,
-      expert: 3,
-    },
-  }
-}
+import { createStatisticalDataPoint } from './utils/createStatisticalDataPoint'
 
 interface GenderDiagramProps {
   participationsWithAssessment: ParticipationWithAssessment[]
