+++ clients/assessment_component/src/assessment/pages/AssessmentStatisticsPage/interfaces/StatisticalDataPoint.ts
@@ -1,0 +1,15 @@
+export interface SkillCounts {
+  novice: number
+  intermediate: number
+  advanced: number
+  expert: number
+}
+
+export interface StatisticalDataPoint {
+  name: string
+  average: number
+  lowerQuartile: number
+  median: number
+  upperQuartile: number
+  counts: SkillCounts
+}
